pyproject.toml
[project]
name = "forf-uas-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0,<3",
    "python-dotenv>=1.2.1,<2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2,<9",
]

[project.scripts]
service = "forf_uas_client:run.main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.pyrefly]
project-include = ["**/*.py*"]


